adding name title substitute feature
This feature is for the problem that somethimes one might want to keep the name clean. For example, one might never use _ for a note but one will use space for the note entry. However, the filename should better not contain space. We could link the entry for example "hello world" to the name "hello_world.notes" (where notes is the suffix)